improve template and fix tests
