+++ {{ cookiecutter.project_slug }}/app/src/main/scala/org/{{ cookiecutter.project_domain }}/{{ cookiecutter.project_subdomain }}/config/ApplicationConfig.scala
@@ -1,0 +1,6 @@
+package org.{{ cookiecutter.project_domain }}.{{ cookiecutter.project_subdomain }}.config
+
+final case class ApplicationConfig(opsServerConfig: OpsServerConfig, restConfig: RestApiConfig)
+
+final case class RestApiConfig(port: Int)
+final case class OpsServerConfig(port: Int)
