+++ {{ cookiecutter.project_slug }}/domain/src/main/scala/org/{{ cookiecutter.project_domain }}/{{ cookiecutter.project_subdomain }}/domain/repositories/Repository.scala
@@ -1,0 +1,10 @@
+package org.{{ cookiecutter.project_domain }}.{{ cookiecutter.project_subdomain }}.domain.repositories
+
+import java.util.UUID
+import scala.language.higherKinds
+
+trait Repository[F[_]] {
+  def get(id: String): F[Option[RepositoryItem]]
+}
+
+final case class RepositoryItem(name: String, id: UUID)
