+++ {{ cookiecutter.project_slug }}/domain/src/test/scala/org/{{ cookiecutter.project_domain }}/{{ cookiecutter.project_subdomain }}/domain/Test.scala
@@ -1,0 +1,1 @@
+package org.{{ cookiecutter.project_domain }}.{{ cookiecutter.project_subdomain }}.domain
