--- {{ cookiecutter.project_slug }}/domain/src/test/scala/org/{{ cookiecutter.project_subdomain }}/com/Test.scala
@@ -1,1 +1,0 @@
-package org.{{ cookiecutter.project_subdomain }}.com
